Hack for Super Micro serialnumber
Super Micro boards all have a value of '1234567890' as their serial number.  This choked rundeck since Java turns this into an INT and the value is cast as a string.  The hack removes the ability to filer on serial numbers within rundeck but does not actually change the underlying fact.